fix(domain/adapters/xml): Fix data-type and tag-name discrepancies
- PaymentSection.Current_number: int -> int | None; '(null)' is valid
  when Automatic_number is off (confirmed from example file)
- GeneralSection.Bet_capital: int -> Decimal; example file stores it
  as '200000.00', not an integer
- _dispatch.py: 'Rgba' -> 'RGBA' tag key; Grisbi 3.x uses all-caps tag

+++ src/gsbparse/domain/sections/general.py
@@ -156,7 +156,7 @@ class GeneralSection(GsbFileSection):
     Add_archive_in_total_balance: bool
     Force_credit_before_debit: bool | None
     Bet_array_column_width: str
-    Bet_capital: int
+    Bet_capital: Decimal
     Bet_currency: int
     Bet_taux_annuel: Decimal
     Bet_index_duree: int
